Moved find_single_file to utils
